test/web-platform-tests/to-upstream/console/methods.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>console methods</title>
<link rel="author" title="Domenic Denicola" href="mailto:d@domenic.me">
<link rel="help" href="https://console.spec.whatwg.org/#console-interface">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
"use strict";
/* eslint-disable no-console */

test(() => {

  const methods = [
    "assert",
    "clear",
    "count",
    "countReset",
    "debug",
    "dir",
    "dirxml",
    "error",
    "info",
    "log",
    "table",
    "trace",
    "warn",
    "group",
    "groupCollapsed",
    "groupEnd",
    "time",
    "timeLog",
    "timeEnd"
  ];

  for (const method of methods) {
    assert_true(typeof console[method] === "function", `${method} must be a function`);
  }

}, "Test that all methods are present and are functions");

</script>
